requirements.txt
git+https://github.com/wazo-pbx/xivo-lib-rest-client.git
git+https://github.com/wazo-pbx/xivo-ctid-ng-client.git@more-applications
git+https://github.com/wazo-pbx/xivo-auth-client.git
git+https://github.com/wazo-pbx/xivo-confd-client.git
websocket-client
urllib3
pyaml
